Move `merge_async_iterators` to `utils`
